+++ task2/cmd/server/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "main/internal/server"
+
+func main() {
+	server.RunServer()
+}
