+++ task2/pkg/model/decode.go
@@ -1,0 +1,9 @@
+package model
+
+type DecodeRequest struct {
+	Base64String string
+}
+
+type DecodeResponse struct {
+	DecodeString string
+}
